major changes on the overall unpacking logic: 1) all callback classes that are intended to do unpacking and architecture inference mush have the architecture name as a sub-directory at workdir, and unpacked files inside 2) callback system: callback classes can return a handler that would be called just before processing the next result, could also set/unset skip flag to decide the operation of the following results 3) implement xerox dlm firmware unpack class && xz decompression 4) various bug fixes

+++ walk.py
@@ -27,6 +27,18 @@ def safe_filemove(src, dst):
         dst += "_"
     shutil.move(src, dst)
 
+def safe_filecopy(src, dst):
+    while os.path.exists(dst):
+        dst += "_"
+    shutil.copy(src, dst)
+
+def safe_mkdir(parent, ndir):
+    np = os.path.join(parent, ndir)
+    while os.path.exists(np):
+        np += "_"
+    os.mkdir(np)
+    return np
+
 def path2name(pstr):
     if pstr[0] == '.':
         pstr = pstr[1:]
@@ -44,7 +56,7 @@ extension_blacklist = [
         '.make', '.js', '.xml', '.html', '.htm', '.css', '.svn-base',
         '.s', '.txt', '.in', '.asm', '.am', '.log', '.pl', '.png', '.jpg',
         '.gif', '.bmp', '.conf', '.texi', '.plo', '.tex', '.man', '.8',
-        '.sgml', '.diff', '.patch', '.txt', '.pdf',
+        '.sgml', '.diff', '.patch', '.txt', '.pdf', '.class', '.jar',
         ]
 
 def interesting_path(pstr):
@@ -77,8 +89,8 @@ def path_flatten(pstr, toplevel=None):
     # remove all sub directories
     subs = os.listdir(pstr)
     for d in subs:
-        if os.path.isdir(d):
-            shutil.rmtree(d)
+        if os.path.isdir(os.path.join(pstr, d)):
+            shutil.rmtree(os.path.join(pstr, d))
 
 def list_files(pstr):
     res = []
@@ -88,6 +100,11 @@ def list_files(pstr):
             if interesting_path(p):
                 res.append(p)
     return res
+
+rootfs_toplevel = [
+        'usr', 'etc', 'opt', 'var',
+        'bin', 'sbin',
+        ]
 
 import binwalk.core.magic
 
@@ -110,7 +127,7 @@ class CALLBACK(object):
     def __exit__(self, typ, val, tb):
         self.cleanup()  # Dtor from derived class
 
-    def update(self, off, size, workdir):
+    def update(self, desc, off, size, workdir):
         pass
 
     def init(self):
@@ -119,8 +136,50 @@ class CALLBACK(object):
     def cleanup(self):
         pass
 
-    def workspace_cleanup(self, workdir):
-        archs = [sub for sub in os.listdir(workdir) if not sub.startswith(LOGGING)]
+    def save(self, dest, data):
+        destdir = os.path.dirname(dest)
+        if not os.path.exists(destdir):
+            os.mkdirs(destdir)
+        while os.path.exists(dest):
+            dest += "_"
+        with open(dest, 'wb') as fd:
+            fd.write(data)
+
+    def rootfs_handler(self, workdir, destdir, unpack_cb):
+        # cleanup workdir
+        #for sub in os.listdir(workdir):
+        #    shutil.rmtree(os.path.join(workdir, sub))
+
+        unpacked_dir = safe_mkdir(workdir, "unpacked")
+        unpack_cb(unpacked_dir)
+
+        # infer arch
+        assumed_archs = []
+        magic = binwalk.core.magic.Magic(include=['.*'])
+        exesig = os.path.join(BINWALK_DIR, 'magic/executables')
+        magic.load(exesig)
+        for root, _, files in os.walk(unpacked_dir):
+            for f in files:
+                if os.path.islink(os.path.join(root, f)):
+                    continue
+                if not interesting_path(f):
+                    continue
+                with open(os.path.join(root, f), 'rb') as fd:
+                    code = fd.read()
+                if magic.scan(binwalk.core.compat.bytes2str(code)):
+                    assumed_archs.append(self.checkasm(code))
+
+        arch = max(assumed_archs, key=assumed_archs.count)
+
+        # copy results
+        if arch:
+            dest = os.path.join(destdir, arch)
+            os.mkdir(dest)
+            safe_filemove(unpacked_dir, os.path.join(dest, os.path.basename(unpacked_dir)))
+            safe_filecopy(self.binfile, os.path.join(dest, os.path.basename(self.binfile)))
+
+    def workspace_cleanup(self, workdir, flatten_dir=False):
+        archs = [sub for sub in os.listdir(workdir) if os.path.isdir(os.path.join(workdir, sub))]
         if not archs:
             return
 
@@ -132,9 +191,10 @@ class CALLBACK(object):
                 shutil.rmtree(os.path.join(workdir, d))
 
         # flatten dir
-        for f in os.listdir(os.path.join(workdir, self.arch)):
-            safe_filemove(os.path.join(workdir, self.arch, f), os.path.join(workdir, f))
-        shutil.rmtree(os.path.join(workdir, self.arch))
+        if flatten_dir:
+            for f in os.listdir(os.path.join(workdir, self.arch)):
+                safe_filemove(os.path.join(workdir, self.arch, f), os.path.join(workdir, f))
+            shutil.rmtree(os.path.join(workdir, self.arch))
 
     def checkasm(self, code):
         self.magic.reset()
@@ -219,7 +279,7 @@ class LZMA_CB(CALLBACK):
             unpacked += buf
         return unpacked, True
 
-    def update(self, off, size, workdir):
+    def update(self, desc, off, size, workdir):
         fd = binwalk.core.common.BlockFile(self.binfile)
         fd.seek(off)
         data = fd.read()
@@ -229,10 +289,47 @@ class LZMA_CB(CALLBACK):
             unpacked, valid = self._try_deflate(data[:5]+'\xff'*8+data[5:])
 
         self.arch = self.checkasm(unpacked)
-        with open(os.path.join(workdir, f"lzma_{self.index}"), 'wb') as fd:
+        if not self.arch:
+            return
+
+        os.mkdir(os.path.join(workdir, self.arch))
+        with open(os.path.join(workdir, self.arch, f"lzma_{self.index}"), 'wb') as fd:
             fd.write(unpacked)
 
         self.index += 1
+
+import shutil
+import tempfile
+class XZ_CB(CALLBACK):
+    def init(self):
+        self.index = 0
+
+    def update(self, desc, off, size, workdir):
+        fd = binwalk.core.common.BlockFile(self.binfile)
+        fd.seek(off)
+        data = fd.read()
+
+        temp_dir = tempfile.mkdtemp("_tmpx")
+        decomp = lzma.LZMADecompressor()
+        stride = 0x10
+        with open(os.path.join(temp_dir, "tmp"), 'wb') as fd:
+            for i in range(0, len(data), stride):
+                try:
+                    unpacked = decomp.decompress(binwalk.core.compat.str2bytes(data[i:i+stride]))
+
+                    fd.write(unpacked)
+                except lzma.LZMAError as e:
+                    break
+                except EOFError as e:
+                    break
+
+        self.index += 1
+
+        Extractor(os.path.join(temp_dir, "tmp"), toplevel=temp_dir).extract(workdir, extra_file_dir=False)
+
+        self.workspace_cleanup(workdir)
+        if not DEBUG:
+            shutil.rmtree(temp_dir)
 
 
 class SQUASHFS_CB(CALLBACK):
@@ -240,7 +337,7 @@ class SQUASHFS_CB(CALLBACK):
         self.index = 0
 
     # TODO: parse squashfs - https://github.com/plougher/squashfs-tools
-    def update(self, off, size, workdir):
+    def update(self, desc, off, size, workdir):
         fd = binwalk.core.common.BlockFile(self.binfile)
         fd.seek(off)
         data = fd.read(size)
@@ -256,7 +353,7 @@ class ZIP_CB(CALLBACK):
     def init(self):
         self.seen_header = False
 
-    def update(self, off, size, workdir):
+    def update(self, desc, off, size, workdir):
         if self.seen_header:
             return
         self.seen_header = True
@@ -267,12 +364,16 @@ class ZIP_CB(CALLBACK):
         fd.close()
 
         temp_dir = tempfile.mkdtemp('_tmpx')
-        with zipfile.ZipFile(io.BytesIO(binwalk.core.compat.str2bytes(data))) as z:
-            for zi in z.infolist():
-                if not zi.is_dir() and interesting_path(zi.filename):
-                    newfn = path2name(zi.filename)
-                    with open(os.path.join(temp_dir, newfn), 'wb') as fd:
-                        fd.write(z.read(zi))
+        try:
+            with zipfile.ZipFile(io.BytesIO(binwalk.core.compat.str2bytes(data))) as z:
+                for zi in z.infolist():
+                    if not zi.is_dir() and interesting_path(zi.filename):
+                        newfn = path2name(zi.filename)
+                        with open(os.path.join(temp_dir, newfn), 'wb') as fd:
+                            fd.write(z.read(zi))
+        except zipfile.BadZipFile as e:
+            #print("Bad Zip File")
+            pass
 
         for f in list_files(temp_dir):
             Extractor(f, toplevel=temp_dir).extract(workdir, extra_file_dir=False)
@@ -287,7 +388,7 @@ class ZIP_CB(CALLBACK):
 import zlib
 import tempfile
 class ZLIB_CB(CALLBACK):
-    def update(self, off, size, workdir):
+    def update(self, desc, off, size, workdir):
         fd = binwalk.core.common.BlockFile(self.binfile)
         fd.seek(off)
         data = fd.read()
@@ -312,7 +413,7 @@ class ZLIB_CB(CALLBACK):
 import tempfile
 import subprocess
 class RAR_CB(CALLBACK):
-    def update(self, off, size, workdir):
+    def update(self, desc, off, size, workdir):
         fd = binwalk.core.common.BlockFile(self.binfile)
         fd.seek(off)
         data = fd.read()
@@ -340,28 +441,26 @@ class RAR_CB(CALLBACK):
 import io
 import gzip
 class GZIP_CB(CALLBACK):
-    def update(self, off, size, workdir):
+    def update(self, desc, off, size, workdir):
         fd = binwalk.core.common.BlockFile(self.binfile)
         fd.seek(off)
         data = fd.read()
 
-        unpacked = b""
-        stride = 0x10
+        stride = 10 * 1024  # let's trust binwalk plugin for the block size for now
         gz = gzip.GzipFile(fileobj=io.BytesIO(binwalk.core.compat.str2bytes(data)))
-        for i in range(0, len(data), stride):
-            try:
-                buf = gz.read(stride)
-            except zlib.error as e:
-                #print("invalid gzip")
-                return
-            except OSError as e:
-                #print("invalid gzip")
-                break
-            unpacked += buf
 
         temp_dir = tempfile.mkdtemp('_tmpx')
         with open(os.path.join(temp_dir, "tmp"), 'wb') as fd:
-            fd.write(unpacked)
+            for i in range(0, len(data), stride):
+                try:
+                    unpacked = gz.read(stride)
+                except zlib.error as e:
+                    #print("invalid gzip")
+                    return
+                except OSError as e:
+                    #print("invalid gzip")
+                    break
+                fd.write(unpacked)
 
         Extractor(os.path.join(temp_dir, "tmp"), toplevel=temp_dir).extract(workdir, extra_file_dir=False)
 
@@ -371,7 +470,7 @@ class GZIP_CB(CALLBACK):
 
 import bz2
 class BZIP2_CB(CALLBACK):
-    def update(self, off, size, workdir):
+    def update(self, desc, off, size, workdir):
         fd = binwalk.core.common.BlockFile(self.binfile)
         fd.seek(off)
         data = fd.read()
@@ -385,13 +484,14 @@ class BZIP2_CB(CALLBACK):
         Extractor(os.path.join(temp_dir, "tmp"), toplevel=temp_dir).extract(workdir, extra_file_dir=False)
 
         self.workspace_cleanup(workdir)
-        # shutil.rmtree(temp_dir)
+        if not DEBUG:
+            shutil.rmtree(temp_dir)
 
 import io
 import tarfile
 import tempfile
 class TAR_CB(CALLBACK):
-    def update(self, off, size, workdir):
+    def update(self, desc, off, size, workdir):
         fd = binwalk.core.common.BlockFile(self.binfile)
         fd.seek(off)
         data = fd.read()
@@ -399,17 +499,34 @@ class TAR_CB(CALLBACK):
         try:
             tar = tarfile.TarFile(fileobj=io.BytesIO(binwalk.core.compat.str2bytes(data)))
         except tarfile.InvalidHeaderError as e:
+            #print("Tar Error: invalid Headder")
             return
 
+        found_fs = False
         temp_dir = tempfile.mkdtemp('_tmpx')
-        for m in tar.getmembers():
-            if m.isfile() and interesting_path(m.name):
-                buf = tar.extractfile(m)
-                with open(os.path.join(temp_dir, path2name(m.name)), 'wb') as fd:
-                    fd.write(buf.read())
+        try:
+            for m in tar:
+                if m.isdir() and m.name in rootfs_toplevel:
+                    found_fs = True
+                    break
+                if m.isfile() and interesting_path(m.name):
+                    buf = tar.extractfile(m)
+                    with open(os.path.join(temp_dir, path2name(m.name)), 'wb') as fd:
+                        fd.write(buf.read())
+        except tarfile.ReadError as e:
+            #print("corrupted tar file")
+            pass
 
-        for f in list_files(temp_dir):
-            Extractor(f, toplevel=temp_dir).extract(workdir, extra_file_dir=False)
+        if found_fs:
+            def unpack_cb(unpackdir):
+                try:
+                    tar.extractall(unpackdir)
+                except tarfile.ReadError as e:
+                    pass
+            self.rootfs_handler(temp_dir, workdir, unpack_cb)
+        else:
+            for f in list_files(temp_dir):
+                Extractor(f, toplevel=temp_dir).extract(workdir, extra_file_dir=False)
 
         self.workspace_cleanup(workdir)
         if not DEBUG:
@@ -417,7 +534,7 @@ class TAR_CB(CALLBACK):
 
 
 class VXWORKS_CB(CALLBACK):
-    def update(self, off, size, workdir):
+    def update(self, desc, off, size, workdir):
         fd = binwalk.core.common.BlockFile(self.binfile)
         fd.seek(off)
         data = fd.read()
@@ -427,15 +544,49 @@ class VXWORKS_CB(CALLBACK):
         #with open(os.path.join(workdir, "vxworks"), 'wb') as fd:
         #    fd.write(binwalk.core.compat.str2bytes(data))
 
+import tempfile
+class XEROXDLM_CB(CALLBACK):
+    def _unpack_data(self, result, caller, workdir):
+        caller.callnext = None
+
+        fd = binwalk.core.common.BlockFile(self.binfile)
+        fd.seek(result.offset)
+        data = fd.read()
+
+        temp_dir = tempfile.mkdtemp("_tmpx")
+        with open(os.path.join(temp_dir, "tmp"), 'wb') as fd:
+            fd.write(binwalk.core.compat.str2bytes(data))
+
+        extractor = Extractor(os.path.join(temp_dir, "tmp"), toplevel=temp_dir)
+        cb = extractor.dispatch_callback(result.description)
+        cb.update(result.description, 0, -1, workdir)
+
+        cb.workspace_cleanup(workdir)
+        self.arch = cb.arch
+
+        if self.arch:
+            caller.assumed_archs.append(self.arch)
+
+    def _skip_header(self, result, caller, workdir):
+        desc = result.description.lower()
+        if desc.startswith("xerox dlm firmware end of header"):
+            caller.callnext = self._unpack_data
+        # Skip the rest
+        caller.skip = True
+
+    def update(self, desc, off, size, workdir):
+        if desc.lower().startswith("xerox dlm firmware start of header"):
+            return self._skip_header
+
 class HTML_CB(CALLBACK):
     def init(self):
         self.counter = 0
 
-    def update(self, off, size, workdir):
+    def update(self, desc, off, size, workdir):
         self.counter += 1
 
 class ELF_CB(CALLBACK):
-    def update(self, off, size, workdir):
+    def update(self, desc, off, size, workdir):
         fd = binwalk.core.common.BlockFile(self.binfile)
         fd.seek(off)
         data = fd.read(0x1000)  # read first page
@@ -443,7 +594,7 @@ class ELF_CB(CALLBACK):
         self.arch = self.checkasm(data)
 
 class LINUXKERN_CB(CALLBACK):
-    def update(self, off, size, workdir):
+    def update(self, desc, off, size, workdir):
         fd = binwalk.core.common.BlockFile(self.binfile)
         fd.seek(0)
         data = fd.read()
@@ -457,8 +608,11 @@ class Extractor(object):
     def __init__(self, binfile, toplevel="/data/firmware/images"):
         self.toplevel = os.path.abspath(os.path.realpath(toplevel))
         self.binfile = os.path.abspath(os.path.realpath(binfile))
+        self.skip = False
+        self.callnext = None
 
         self.lzma_cb = LZMA_CB(self.binfile)
+        self.xz_cb = XZ_CB(self.binfile)
         self.squashfs_cb = SQUASHFS_CB(self.binfile)
         self.zip_cb = ZIP_CB(self.binfile)
         self.zlib_cb = ZLIB_CB(self.binfile)
@@ -470,14 +624,19 @@ class Extractor(object):
         self.bzip2_cb = BZIP2_CB(self.binfile)
         self.tar_cb = TAR_CB(self.binfile)
         self.linuxkern_cb = LINUXKERN_CB(self.binfile)
+        self.xeroxdlm_cb = XEROXDLM_CB(self.binfile)
 
     def dispatch_callback(self, desc):
         if desc.lower().startswith("lzma compressed data"):
             return self.lzma_cb
+        elif desc.lower().startswith("xz compressed data"):
+            return self.xz_cb
         elif desc.lower().startswith("squashfs filesystem"):
             return self.squashfs_cb
         elif desc.lower().startswith("zip archive data"):
-            return self.zip_cb
+            match = re.search(", name: (.+?)$",desc.lower())
+            if not match or interesting_path(match.group(1)):
+                return self.zip_cb
         elif desc.lower().startswith("end of zip archive"):
             self.zip_cb.reset()
             # fall through
@@ -494,8 +653,6 @@ class Extractor(object):
             match = re.search("has original file name: \"(.+?)\",",desc.lower())
             if not match or interesting_path(match.group(1)):
                 return self.gzip_cb
-            else:
-                return self.html_cb # re-use html indicator
         elif desc.lower().startswith("elf, "):
             return self.elf_cb
         elif desc.lower().startswith("bzip2 compressed data"):
@@ -504,6 +661,8 @@ class Extractor(object):
             return self.tar_cb
         elif desc.lower().startswith("linux kernel version"):
             return self.linuxkern_cb
+        elif desc.lower().startswith("xerox dlm firmware start of header"):
+            return self.xeroxdlm_cb
 
         # failsafe
         return CALLBACK(self.binfile)
@@ -517,33 +676,38 @@ class Extractor(object):
             sigmod = executed_mods[0]
             assert(isinstance(sigmod, binwalk.modules.Signature))
 
-            assumed_archs = []
+            self.assumed_archs = []
             for result in sigmod.results:
-                if result.valid:
+                if self.callnext:
+                    self.callnext(result, self, temp_dir)
+                if not self.skip and result.valid:
                     cb = self.dispatch_callback(result.description)
-                    cb.update(result.offset, result.size, temp_dir)
+                    self.callnext = cb.update(result.description, result.offset, result.size, temp_dir)
                     if cb.arch:
-                        assumed_archs.append(cb.arch)
+                        self.assumed_archs.append(cb.arch)
 
         # special case, if we seen a lot html header/footer, that should be the firmware
-        if self.html_cb.counter >= len(sigmod.results)/2:
+        if self.html_cb.counter > 0:
             with open(self.binfile, 'rb') as fd:
                 check_arch = CALLBACK(self.binfile).checkasm(fd.read())
             if check_arch:
                 for i in range(self.html_cb.counter):
-                    assumed_archs.append(check_arch)
-                shutil.copy(self.binfile, temp_dir)
-        if not assumed_archs:
+                    self.assumed_archs.append(check_arch)
+                archpath = os.path.join(temp_dir, check_arch)
+                if not os.path.exists(archpath):
+                    os.mkdir(archpath)
+                shutil.copy(self.binfile, archpath)
+        if not self.assumed_archs:
             return False
 
-        arch = max(assumed_archs, key=assumed_archs.count)
+        arch = max(self.assumed_archs, key=self.assumed_archs.count)
         rel_path = os.path.relpath(os.path.dirname(self.binfile), self.toplevel)
         dest_path = os.path.abspath(os.path.join(workdir, arch, rel_path))
         if extra_file_dir:
             dest_path = os.path.join(dest_path, os.path.basename(self.binfile))
         os.makedirs(dest_path, exist_ok=True)
-        for fn in os.listdir(temp_dir):
-            safe_filemove(os.path.join(temp_dir, fn), os.path.join(dest_path, fn))
+        for fn in os.listdir(os.path.join(temp_dir, arch)):
+            safe_filemove(os.path.join(temp_dir, arch, fn), os.path.join(dest_path, fn))
 
         return True
